Reduce number of checks when waiting for SPI completion
Fix GPE1 timing fw sensor

CQ: SW396887
Reviewed-on: http://ralgit01.raleigh.ibm.com/gerrit1/43706
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>

+++ src/common/gpe_util.h
@@ -33,7 +33,7 @@
 
 void gpe_set_ffdc(GpeErrorStruct *o_error, uint32_t i_addr, uint32_t i_rc, uint64_t i_ffdc);
 
-int wait_spi_completion(GpeErrorStruct *error, uint32_t reg, uint8_t timeout);
+int wait_spi_completion(GpeErrorStruct *error, uint32_t reg, uint32_t timeout);
 
 // Read decrementer register
 #define MFDEC(reg_var) \
